ai-banking-agents/requirements.txt
fastapi
uvicorn
SQLAlchemy
mysqlclient
python-jose[cryptography]
passlib[bcrypt]
requests
python-dotenv
pillow
easyocr
langchain
langgraph
